+++ package.json
@@ -74,14 +74,15 @@
 	},
 	"dependencies": {
 		"@hookform/resolvers": "^3.10.0",
-		"@uniteam31/uni-shared-ui": "^0.0.13",
 		"@uniteam31/uni-shared-toolkit": "0.0.1",
+		"@uniteam31/uni-shared-types": "0.0.2",
+		"@uniteam31/uni-shared-ui": "^0.0.13",
+		"axios": "1.7.9",
 		"classnames": "^2.5.1",
 		"dotenv-defaults": "^5.0.2",
 		"immer": "^10.1.1",
 		"react": "18.3.1",
 		"react-dom": "18.3.1",
-		"axios": "1.7.9",
 		"react-hook-form": "^7.53.2",
 		"react-router": "7.0.2",
 		"react-router-dom": "6.22.1",
