--- src/entities/Note/model/types/note.ts
@@ -1,15 +1,0 @@
-import { TMeta } from 'shared/types';
-
-/** Поля у формы заметки */
-export type TNoteFormFields = {
-	title?: string;
-	text?: string;
-};
-
-/** Данные заметки */
-export type TNodeData = TNoteFormFields & {
-	author: string;
-};
-
-/** Целый экземпляр */
-export interface INote extends TNodeData, TMeta {}
