--- src/entities/Note/ui/Item/Item.module.scss
@@ -1,16 +1,0 @@
-.Item {
-    flex: 1;
-    padding: 30px 0 30px 30px;
-    position: relative;
-
-    .trashIcon {
-        cursor: pointer;
-        position: absolute;
-        right: 20px;
-        top: 20px;
-    }
-
-    .title {
-        margin-bottom: 10px;
-    }
-}
