+++ src/shared/api/index.ts
@@ -1,1 +1,2 @@
+/** Common API */
 export { axiosInstance } from '@uniteam31/uni-shared-toolkit';
