+++ tsconfig.json
@@ -20,5 +20,10 @@
 		"compilerOptions": {
 			"module": "CommonJS"
 		}
-	}
+	},
+	"include": ["src/**/*"],
+	"exclude": [
+		"node_modules",
+		"webpack.config.ts"
+	]
 }
